Disable download button for over 20k test results
