+++ frontend/src/app/testResults/types.d.ts
@@ -17,3 +17,12 @@ interface PxpMultiplexResult extends DiseaseName {
 
 type MultiplexResult = SRMultiplexResult | PxpMultiplexResult;
 type MultiplexResults = MultiplexResult[];
+
+type FilterParams = {
+  patientId?: string | null;
+  startDate?: string | null;
+  endDate?: string | null;
+  role?: string | null;
+  result?: string | null;
+  filterFacilityId?: string | null;
+};
